requirements.txt
python-telegram-bot==21.7
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
numpy==2.0.0